Add Scalar and List options for Char #471
